+++ src/app/components/reports/logi/logiController.js
@@ -24,7 +24,43 @@ angular.module('liveopsConfigPanel')
             EmbeddedReporting.get('logiContainer').iframe) {
           clearInterval(loadedCheck);
           EmbeddedReporting.get('logiContainer').iframe.addEventListener('load', function(){
-            console.warn('logiContainer iframe has finished loading');
+            console.warn('logiContainer iframe has finished loading ', EmbeddedReporting.get('logiContainer').iframe.id);
+            var iframeId = '#' + EmbeddedReporting.get('logiContainer').iframe.id;
+            $(iframeId).css('height', 'calc(100vh - 54px)');
+            // parse out just the domain with no subdomain
+            var domainOnly;
+            // if there is a domain suffix, split it up to grab the domain only
+            if ($window.location.hostname.split('.').length > 2) {
+              domainOnly = $window.location.hostname.split('.').slice(1).join('.');
+            } else {
+              // otherwise (as in the case of 'localhost', just get the hostname)
+              domainOnly = $window.location.hostname;
+            }
+            document.domain = domainOnly;
+            $scope.iframeLoaded = true;
+          });
+        }
+      }, 200);
+
+      var loadedCheck2 = setInterval(function () {
+        if (window.EmbeddedReporting && 
+            EmbeddedReporting.get('ssmContainer') &&
+            EmbeddedReporting.get('ssmContainer').iframe) {
+          clearInterval(loadedCheck2);
+          EmbeddedReporting.get('ssmContainer').iframe.addEventListener('load', function(){
+            console.warn('ssmContainer iframe has finished loading ', EmbeddedReporting.get('ssmContainer').iframe.id);
+            var iframeId = '#' + EmbeddedReporting.get('ssmContainer').iframe.id;
+            $(iframeId).css('height', 'calc(100vh - 54px)');
+            // parse out just the domain with no subdomain
+            var domainOnly;
+            // if there is a domain suffix, split it up to grab the domain only
+            if ($window.location.hostname.split('.').length > 2) {
+              domainOnly = $window.location.hostname.split('.').slice(1).join('.');
+            } else {
+              // otherwise (as in the case of 'localhost', just get the hostname)
+              domainOnly = $window.location.hostname;
+            }
+            document.domain = domainOnly;
             $scope.iframeLoaded = true;
           });
         }
@@ -38,7 +74,7 @@ angular.module('liveopsConfigPanel')
             applicationUrl: response.data.logiBaseUrl,
             linkParams: {'rdSecurekey': response.data.secureToken, 'tenantID': Session.tenant.tenantId},
             report: 'Common.Bookmarks',
-            autoSizing: 'all',
+            autoSizing: 'width',
           });
         })
         .catch(function(err) {
@@ -51,7 +87,7 @@ angular.module('liveopsConfigPanel')
             applicationUrl: response.data.logiBaseUrl,
             linkParams: {'rdSecurekey': response.data.secureToken, 'tenantID': Session.tenant.tenantId},
             report: 'InfoGo.goHome',
-            autoSizing: 'all',
+            autoSizing: 'width',
           });
         })
         .catch(function(err) {